feat: restructure backend app initialization and middleware setup; remove unused imports and enhance error handling
