+++ backend/src/models/message.model.js
@@ -18,9 +18,6 @@ const messageSchema = new mongoose.Schema(
     text: {
       type: String,
     },
-    image: {
-      type: String,
-    },
     reactions: {
       type: [{
         userId: {
